backend/requirements.txt
asgiref==3.8.1
celery==5.5.0rc5
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.12.1
djangorestframework~=3.15.1
django-allauth~=0.61.1
django-cors-headers~=4.3.1
django-extensions~=3.2.3
dj-rest-auth~=6.0.0
pyopenssl~=24.1.0
redis~=5.3.0b5
Werkzeug~=3.0.3
web3~=7.9.0